LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-pseudo-open-expected.txt


one
two

PASS select should support :open pseudo selector.
FAIL select :open and :not(:open) should invalidate correctly. assert_equals: The style rules from :not(:open) should apply when the select is opened and closed again. expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
